@@ -28,13 +28,13 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "4.8.*"
+        "phpunit/phpunit": "9.*"
     },
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.3",
         "g4/constants": "*",
         "g4/profiler": ">=1.10.0",
-        "laminas/laminas-http": "2.9.*",
+        "laminas/laminas-http": "2.14.*",
         "g4/value-object": "*"
     }
 }
